+++ R/data-columns.R
@@ -1,0 +1,13 @@
+#' kospi and kosdaq master columns
+#'
+#' These datassets contain both English and Korean column names.
+#'
+#' @format data.frame
+#' @examples
+#' kospi_master_columns
+#' kosdaq_master_columns
+"kospi_master_columns"
+
+#' @rdname kospi_master_columns
+#' @format NULL
+"kosdaq_master_columns"