+++ R/data-kospi.R
@@ -1,8 +1,0 @@
-#' Kospi master columns
-#'
-#' This dataset contains both English and Korean column names
-#'
-#' @format data.frame
-#' @examples
-#' kospi_master_columns
-"kospi_master_columns"